requirements.txt
fastmcp>=2.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.4.2
